Adding Taming Prototype
If the player presses the 'T' key when inside the 'animal' trigger volume, the 'animal' will then follow the player, remaining a minimum distance away at all times.
Added two scripts AnimalTriggerBox.cpp to determine when the player is inside or outside of the 'animal' trigger volume, and TameAnimal.cpp (attached to the 'animal' cube) which handles the player keypress and updates the position of the animal so it appears to follow the player.