Taming Prototype Update #1
Animal now faces player at all times when tamed. Animal has a sight cone (defined in code), and if the player enters this cone, the animal turns away from the player and flees to a set distance. The angle of vision and the fleeing distance can both be changed in the header file of the new AnimalSight script.