Taming Prototype Update #2
Animal now has an exhaustion variable, that increases as it flees from the player and decreases, whenever it is stationary. If the exhaustion value reaches a preset maximum value, the animal with become exhausted and will then be tameable. If the player doesn't tame the animal quickly enough and the exhaustion level returns to 0, then the animal will begin to flee again.
If the player moves a certain preset distance from the animal at any time while the animal is alerted or exhausted, the hunt will be abandoned, and the animal will return to its default unalerted state.
If the player untames the animal at any point, it will immediately be alerted, regardless of facing direction and will begin to flee.
Added a limit to how far the animal can see, which is defined by a variable in the AnimalSight header file.
Tidied up code a lot, and introduced an AnimalMotion class that contains most of the functions related to AnimalMotion, and it contains a lot of variables that are shared by both AnimalSight and TameAnimal. Now AnimalSight and TameAnimal can both see the attached instance of AnimalMotion, and both use it, but neither of them require any information from one another.