Riding Prototype & some changes to Taming Prototype setup
Riding Prototype:
Added Riding Prototype (added RideAnimal.cpp which contains the relevant code).
Animal has been replaced with a purple male human figure, and player has been replaced with a white version of the same model.
When the player presses "R" within the interaction trigger volume, the playercontroller un-possesses the player character and possesses the animal character instead. The animation state for the player character is changed from standing to riding. The player character's position and rotation are constantly updated to a forward facing position just above the animal's shoulders.The user can then control the animal with the player riding on top.
When the player presses "R" again, the player dismounts, and the playercontroller possesses the player character again. The player is always placed a short distance to the right of the animal, with its animation state returned to standing.

Taming Prototype:
Changed the animal from a pawn to a playable character, and adjusted the Taming Prototype code a little to accomodate for this.

To Do:
Need to tidy up RideAnimal.cpp.